finalize token embeddings
