+++ flair/embeddings/__init__.py
@@ -37,6 +37,7 @@ from .legacy import (
     CharLMEmbeddings,
     DocumentLSTMEmbeddings,
     DocumentMeanEmbeddings,
+    ELMoEmbeddings,
     ELMoTransformerEmbeddings,
     OpenAIGPT2Embeddings,
     OpenAIGPTEmbeddings,
@@ -51,7 +52,6 @@ from .token import (
     BPEmbSerializable,
     BytePairEmbeddings,
     CharacterEmbeddings,
-    ELMoEmbeddings,
     FastTextEmbeddings,
     FlairEmbeddings,
     HashEmbeddings,
